requirements.txt
pyyaml>=6.0
tqdm>=4.66.0
colorama>=0.4.6

# Web Dashboard
Flask>=3.0.0
flask-cors>=4.0.0
# Optional: IP-Adapter support (uncomment when ready)
# ML / Generation
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
scipy>=1.10.0

# Optional: Free AI enhancement (no API key needed)
# Install Ollama for local AI: https://ollama.ai
requests>=2.31.0  # For optional Ollama support


# Utilities
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.66.0
colorama>=0.4.6
python-dotenv>=1.0.0  # For .env config

# Development & Testing
pytest>=7.4.0
